Find instances where pinning memory is not seen
I wasn't able to capture all the cases with the original rule. More concretely, I wasn't able to write a rule that would capture instances when pin_memory isn't set at all and when it is set to False. As I was doing more research into this issue, it turns out there are a lot of cases where automatic memory pinning is not possible, so I decided to flag cases where the argument isn't set. That way, the rule will find instances where the implications of automatically pinning memory were not considered.